+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Bugfixes
 
  - Fix saving settings of Conditional Timers ([#470](https://github.com/kizniche/mycodo/issues/470))
+ - Fix Command PWM output use in PIDs ([#469](https://github.com/kizniche/mycodo/issues/469))
 
 
 ## 6.1.0 (2018-05-02)